rateOfAchievement Double -> int 변경

+++ src/main/java/team/crowdee/domain/dto/ThumbNailDTO.java
@@ -16,7 +16,7 @@ public class ThumbNailDTO {
     private String title;
     private String summary;
     private String tag;
-    private Double rateOfAchievement;
+    private int rateOfAchievement;
     private int goalFundraising;
     private String category;
     private int restDate;
